refactor(login): separate the google auth component
