+++ src/components/Button/GoogleSignInButton.tsx
@@ -1,0 +1,36 @@
+import React from "react";
+import { useOAuth } from "@clerk/clerk-expo";
+import * as Linking from "expo-linking";
+import GoogleButton from "@/components/Button/GoogleButton";
+
+const GoogleSignInButton = ({ isLoading, setIsLoading }) => {
+  const { startOAuthFlow } = useOAuth({ strategy: "oauth_google" });
+
+  const onPress = React.useCallback(async () => {
+    setIsLoading(true);
+    try {
+      const { createdSessionId, setActive } = await startOAuthFlow({
+        redirectUrl: Linking.createURL("/(tabs)/home", { scheme: "savy" }),
+      });
+
+      if (createdSessionId) {
+        setActive!({ session: createdSessionId });
+      }
+    } catch (err) {
+      console.error("OAuth error", err);
+    } finally {
+      setIsLoading(false);
+    }
+  }, [startOAuthFlow, setIsLoading]);
+
+  return (
+    <GoogleButton
+      title="Sign in with Google"
+      onPress={onPress}
+      containerStyles="mx-4 py-4"
+      isLoading={isLoading}
+    />
+  );
+};
+
+export default GoogleSignInButton;
